分离事件和图片 + API 文档 + 环境变量
