+++ .env
@@ -1,0 +1,6 @@
+ADMIN_PWD=kxpage-password
+IMAGE_STORE=./assets/images
+DATABASE_PATH=./data/database.db
+IP_ADDRESS=127.0.0.1
+PORT=8000
+PUBLIC_PREFIX=/api/v1
