+++ Cargo.toml
@@ -8,6 +8,7 @@ anyhow = "1.0.102"
 axum = "0.8.8"
 base64 = "0.22.1"
 chrono = "0.4.44"
+dotenv = "0.15.0"
 lazy_static = "1.5.0"
 prost = "0.14.3"
 rusqlite = "0.38.0"
